+++ ColorfulPieces/ComfyLib/Commands/ComfyCommand.cs
@@ -5,6 +5,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
+using UnityEngine;
+
 [AttributeUsage(AttributeTargets.Method)]
 public sealed class ComfyCommand : Attribute {
   // ...
@@ -25,7 +27,9 @@ public static class ComfyCommandUtils {
   }
 
   static void UpdateCommandLists() {
-    foreach (Terminal terminal in UnityEngine.Object.FindObjectsOfType<Terminal>(includeInactive: true)) {
+    foreach (
+        Terminal terminal in
+            UnityEngine.Object.FindObjectsByType<Terminal>(FindObjectsInactive.Include, FindObjectsSortMode.None)) {
       terminal.updateCommandList();
     }
   }
@@ -40,8 +44,8 @@ public static class ComfyCommandUtils {
 
   static IEnumerable<Terminal.ConsoleCommand> RegisterCommands(Assembly assembly) {
     return assembly.GetTypes()
-        .SelectMany(type => type.GetMethods(BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic))
-        .Where(method => method.GetCustomAttributes(typeof(ComfyCommand), inherit: false).Length > 0)
+        .SelectMany(static type => type.GetMethods(BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic))
+        .Where(static method => method.GetCustomAttributes(typeof(ComfyCommand), inherit: false).Length > 0)
         .SelectMany(RegisterCommands);
   }